docs: minor README updates
The previous "brew install" command performed a "brew tap" and "brew
install" in  one line. That's nice, but the "brew tap" is obscured, and if
you try to use the same command to install it a second time, it will fail
(because the formulae has already been tapped). Thus, I decided to split
up the install commands to make it more clear what was happening, and
prevent confusing reinstall issues.